Tests/LibWeb/Text/expected/ServiceWorker/service-worker-container-cross-realm.txt

frame has serviceWorker: true
ready promise uses iframe Promise: true
ready promise uses top Promise: false
ready promise cache is stable: true
getRegistration promise uses iframe Promise: true
getRegistration promise uses top Promise: false
register promise uses iframe Promise: true
register promise uses top Promise: false
